Feat: 보스전 완료시 룬의 색이 변하도록 함
MainRoom에 Anger, Fear, Humiliation, Sorrow, Regret 속성에 해당하는 Rune를 배치하여 해당 속성 보스전을 완료시에 룬의 색깔이 변하도록 함
