+++ EPOH/Assets/Scripts/ActivateRune.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5fe6bc71f9811aa43b89960210ab0ae9
+guid: 70c9c67db28b30c45a5d8123382c8bbb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
